TK-03 - Added structure changes for the summary blocks of Winnability item menu
