+++ src/app/account/account-details/account-details.component.scss
@@ -99,7 +99,7 @@
       gap: 1rem;
 
       .details-title {
-        margin-block: 0 1.5rem;
+        margin: 0;
         font-size: 3rem;
         font-weight: 400;
       }
@@ -108,11 +108,13 @@
         display: flex;
         gap: 0.5rem;
 
-        > div {
+        .summary-item {
           flex: 1;
           display: flex;
           flex-direction: column;
-          padding: 1.25rem;
+          gap: 0.25rem;
+          padding-block: 1rem;
+          padding-inline: 1.25rem;
           background-color: $app-sub-card-bg-color;
           border-radius: 1.875rem;
           color: $app-text-color;
@@ -126,6 +128,28 @@
             display: flex;
             flex-direction: column;
             gap: 0.25rem;
+
+            &.overall-score {
+              align-items: center;
+              flex-direction: row;
+              gap: 0.5rem;
+
+              .title {
+                margin: 0;
+                font-size: 3rem;
+                font-weight: 500;
+              }
+            }
+
+            .dot {
+              width: 0.625rem;
+              height: 0.625rem;
+              border-radius: 50%;
+
+              &.blue {
+                background-color: $app-main-blue-color;
+              }
+            }
           }
         }
       }
